requirements.txt
# AI Finance Assistant - Python Dependencies
# Install: pip install -r requirements.txt

# Core Web Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Data Processing & Validation
pydantic>=2.0
pandas>=2.0

# LLM & AI
openai>=1.3.0
anthropic>=0.7.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langgraph>=0.0.1

# Vector Database
pinecone>=3.0

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0

# Market Data
yfinance==0.2.28

# Token Counting
tiktoken>=0.5.0

# Configuration & Secrets
python-dotenv>=1.0.0

# Utilities
pytz>=2023.3

# Testing (development only)
pytest>=7.4.0
pytest-asyncio>=0.21.0
